Finalizadas atividades pedidas em aula

+++ AulasMJV/src/myapp/pedidos/PedidoItem.java
@@ -5,7 +5,7 @@ import myapp.cadastros.Produto;
 public class PedidoItem {
 	
 	private Integer id;
-	private Double quantidade;
+	private Integer quantidade;
 	private Double valorVenda;
 	private Double valorTotal;
 	private Produto produto;
@@ -26,11 +26,11 @@ public class PedidoItem {
 		this.id = id;
 	}
 	
-	public Double getQuantidade() {
+	public Integer getQuantidade() {
 		return quantidade;
 	}
 	
-	public void setQuantidade(Double quantidade) {
+	public void setQuantidade(Integer quantidade) {
 		this.quantidade = quantidade;
 	}
